nuget

+++ RECmd/Program.cs
@@ -2410,7 +2410,7 @@ internal class Program
             var lines = File.ReadLines(bn).ToList();
             var fileContents = bn.ReadAllText();
 
-            var badLine = lines[se.Start.Line - 1];
+            var badLine = lines[(int)se.Start.Line - 1];
 
             Console.WriteLine();
             Log.Fatal("Bad line (or close to it) {BadLine} has invalid data at column {Column}",badLine,se.Start.Column);
